+++ changelog.md
@@ -83,7 +83,8 @@
 - Removed Clair de Lune Music Disc from V2s Loot Table
 - Fixed Wing Customization Screen messing up FoV Effect Scale Setting
 - Progression Unlock Items now don't get consumed upon use while in creative
-- Progression Unlock Itemsn now use Ultracraft Titles instead of vanilla ones
+- Progression Unlock Items now use Ultracraft Titles instead of Vanilla ones
+- Knuckleblaster Unlcok Item is no longer Stackable
 ## Edit Mode Changes
 - Rooms can now be Parented to Rooms (yay, recursion)
   - Added `/edit config recursiveRooms`, which toggles whether Rooms you place while having another Room focused, will be parented to the selected Room or just be placed as new Root.